requirements.txt
python-telegram-bot==20.4
google-api-python-client==2.97.0
python-dotenv==1.0.0